Rewrite UploadDropZone to ts
